html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-location-replace-set-src.html

<!DOCTYPE html>
<title>Navigating iframe loading='lazy' and then setting src: location.replace</title>
<iframe data-src="support/blank.htm?nav" loading="lazy"></iframe>
<script>
const iframe = document.querySelector('iframe');
iframe.contentWindow.location.replace(iframe.dataset.src);
// Setting src invokes https://html.spec.whatwg.org/multipage/iframe-embed-object.html#process-the-iframe-attributes
iframe.src = 'support/blank.htm?src';
</script>
<!-- Loading testharness.js here is intentional to reproduce a bug in WebKit. -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
onload = () => {
  step_timeout(() => {
    assert_equals(iframe.contentWindow.location.href, new URL("support/blank.htm?src", location.href).href);
    done();
  }, 1000);
};
</script>
<!-- Delay the load event for the same URLs as the potential lazy iframe navigations
     to reduce unstable test results. -->
<iframe src="support/blank.htm?src"></iframe>
<iframe src="support/blank.htm?nav"></iframe>
